feat: playlist 생성

+++ backend/src/main/java/com/ssafy/backend/mysql/entity/PlaylistTrack.java
@@ -25,7 +25,7 @@ public class PlaylistTrack extends Common{
     private Playlist playlist;
 
     @Column(name = "play_order")
-    private Double playOrder;
+    private Integer playOrder;
 
 
 }
